requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-property==0.2.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
deprecation==2.1.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
packaging==24.0
psycopg==3.1.19
psycopg-binary==3.1.19
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.3
pydantic-partial==0.5.5
pydantic-settings==2.3.1
pydantic_core==2.18.4
Pygments==2.18.0
python-dotenv==1.0.1
python-jose==3.3.0
python-keycloak==4.1.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
